Allow add() method to accept multiple modules.
The add() method can now accept a "modules" (plural) argument
that is a hasref containing ModuleName => ModuleVersion pairs.
This can be used as an alternative to (but not in conjunction
with) the "module" (singular) and version arguments.

In practice, any non-trivial distribution is going to contain
multiple modules.  And those modules could all have different
version numbers.  But the old "module" (singular) and "version"
interface did not support that reality.  In fact, mcpi[1]
worked around this limitation by calling add() for each
module in the distribution.  Not only did this seem odd,
it causes the tar.gz file to be needlessly copied several
times.

I would even sugges that the old "module" (singular) and
"version" interface sould be deprecated.

I added test cases to cover the new interface.  Also added
test cases to cover the argument validation logic for the
add() method.

And I updated the POD accordingly.