Improvements on the add() interface.
You may now specify the version argument with the modules (plural)
argument.  This serves as the default version of any modules
that you did not specify an explicit version number for.  This
is convenient for module authors (like me) who assign the same
version number to all modules in a particular release.