The --module option is now awesome.
The --module option for mcpi[1] now takes MODULE_NAME=MODULE_VERSION
arguments and can be repeated for multiple modules.  If you do not
specify a MODULE_VERSION for any given MODULE_NAME, then the value
of the --version option is used as the default.  Here are some
examples:

  # Adds Foo-1.2 and Bar-2.3
  mcpi --add --module Foo=1.2 Bar=2.3 --author ME --file Distro.tar.gz

  # Adds Foo-1.2 and Bar-4.1
  mcpi --add --module Foo=1.2 Bar --version 4.1 --author ME --fle Distro.tar.gz

  # Adds Foo-5.2 and Bar-5.2
  mcpi --add --module Foo --module Bar --version 5.2 --author ME --file Distro.tar.gz

Notice that this is still perfectly compatible with the old
style of using --module and --version to add a single module.